MusicHandling/PythonAndMidi/requirements.txt
mido==1.3.3
packaging==25.0
pylive==0.4.0
python-osc==1.9.3
